complementations to services implementations

+++ src/main/java/com/edanrh/apiong/service/impl/CityServiceImpl.java
@@ -2,10 +2,18 @@ package com.edanrh.apiong.service.impl;
 
 import com.edanrh.apiong.dto.CityDTO;
 import com.edanrh.apiong.dto.converts.CityDTOConvert;
+import com.edanrh.apiong.exceptions.ContentNullException;
+import com.edanrh.apiong.exceptions.DuplicateCreationException;
+import com.edanrh.apiong.exceptions.NotFoundException;
+import com.edanrh.apiong.exceptions.ReferencedEntityException;
 import com.edanrh.apiong.repository.CityRepository;
+import com.edanrh.apiong.repository.HeadquarterRepository;
+import com.edanrh.apiong.repository.ShelterRepository;
 import com.edanrh.apiong.repository.entities.City;
+import com.edanrh.apiong.repository.entities.Headquarter;
 import com.edanrh.apiong.service.CityService;
 import lombok.AllArgsConstructor;
+import org.springframework.http.HttpStatus;
 import org.springframework.stereotype.Service;
 
 import java.util.ArrayList;
@@ -17,47 +25,62 @@ import java.util.Optional;
 public class CityServiceImpl implements CityService {
 
     private CityRepository cityRepository;
+    private HeadquarterRepository headquarterRepository;
+    private ShelterRepository shelterRepository;
     private CityDTOConvert dtoConvert;
 
     @Override
-    public List<CityDTO> findAll() {
+    public List<CityDTO> findAll() throws ContentNullException {
         List<City> cities = (List<City>) cityRepository.findAll();
-        List<CityDTO> citiesdto = new ArrayList<>();
-        for (City city : cities) {
-            citiesdto.add(dtoConvert.toDTO(city));
+        if (cities.isEmpty()){
+            throw new ContentNullException("code", "There isn't cities", HttpStatus.NO_CONTENT);
+        }else {
+            List<CityDTO> citiesDto = new ArrayList<>();
+            for (City city : cities) {
+                citiesDto.add(dtoConvert.toDTO(city));
+            }
+            return citiesDto;
         }
-        return citiesdto;
     }
 
     @Override
-    public CityDTO save(CityDTO cityDTO) {
+    public CityDTO save(CityDTO cityDTO) throws DuplicateCreationException {
         Optional<City> city = cityRepository.findByName(cityDTO.getName());
-        if (city.isEmpty()) {
+        if (city.isPresent()){
+            throw new DuplicateCreationException("code", "City already exists", HttpStatus.CONFLICT);
+        }else {
             City saved = cityRepository.save(dtoConvert.toEntity(cityDTO));
             return dtoConvert.toDTO(saved);
         }
-        return null;
     }
 
     @Override
-    public boolean edit(CityDTO cityDTO) {
+    public boolean edit(CityDTO cityDTO) throws NotFoundException {
         Optional<City> city = cityRepository.findByName(cityDTO.getName());
         if (city.isPresent()) {
             city.get().setName(cityDTO.getName());
             city.get().setDepartment(cityDTO.getDepartment());
             cityRepository.save(city.get());
             return true;
+        }else {
+            throw new NotFoundException("code", "city invalid, don't exist", HttpStatus.NOT_FOUND);
         }
-        return false;
     }
 
     @Override
-    public boolean delete(CityDTO cityDTO) {
+    public boolean delete(CityDTO cityDTO) throws NotFoundException, ReferencedEntityException {
         Optional<City> city = cityRepository.findByName(cityDTO.getName());
         if (city.isPresent()) {
-            cityRepository.delete(city.get());
-            return true;
+            if (shelterRepository.existsByCityName(cityDTO.getName())){
+                throw new ReferencedEntityException("code", "City has shelters linked", HttpStatus.CONFLICT);
+            }else if (headquarterRepository.existsByCityName(cityDTO.getName())){
+                throw new ReferencedEntityException("code", "City has headquarters linked", HttpStatus.CONFLICT);
+            }else {
+                cityRepository.delete(city.get());
+                return true;
+            }
+        }else {
+            throw new NotFoundException("code", "city invalid, don't exist", HttpStatus.NOT_FOUND);
         }
-        return false;
     }
 }
